feat: 🎸 setup NodeCrud implementation
